Move FinalizeProposalError to error mod
